Project dnode should be protected by local MAC
This patch corrects a small security issue with 9c5167d. When the
project dnode was added to the objset_phys_t, it was not included
in the local MAC for cryptographic protection, allowing an attacker
to modify this data without the consent of the key holder. This
patch does represent an on-disk format change for anyone using
project dnodes on an encrypted dataset

Signed-off-by: Tom Caputi <tcaputi@datto.com>
Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>

Change-Id: Id7f876fc4312eff912a8b7856b977f4e95a90f63